==> scRNA_seq_scripts/08_cpdb_ped_healthy.py <==
import pandas as pd
import sys
import os
import anndata


pd.set_option('display.max_columns', 100)

cpdb_file_path = '/cluster/home/t117652uhn/liver_ped_map/data/cellphonedb.zip'
meta_file_path = '/cluster/projects/macparland/RE/PediatricAdult/processed_data/Ped_healthy_meta.csv'
counts_file_path = '/cluster/projects/macparland/RE/PediatricAdult/processed_data/ped_healthy.h5ad'
microenvs_file_path = '/cluster/projects/macparland/RE/PediatricAdult/processed_data/Ped_healthy_microenv.csv'
out_path = '/cluster/projects/macparland/RE/PediatricAdult/processed_data'



## significance analysis
from cellphonedb.src.core.methods import cpdb_statistical_analysis_method

deconvoluted, means, pvalues, significant_means = cpdb_statistical_analysis_method.call(
    cpdb_file_path = cpdb_file_path,                 # mandatory: CellPhoneDB database zip file.
    meta_file_path = meta_file_path,                 # mandatory: tsv file defining barcodes to cell label.
    counts_file_path = counts_file_path,             # mandatory: normalized count matrix.
    counts_data = 'hgnc_symbol',                     # defines the gene annotation in counts matrix.
    microenvs_file_path = microenvs_file_path,       # optional (default: None): defines cells per microenvironment.
    iterations = 1000,                               # denotes the number of shufflings performed in the analysis.
    threshold = 0.1,                                 # defines the min % of cells expressing a gene for this to be employed in the analysis.
    threads = 4,                                     # number of threads to use in the analysis.
    debug_seed = 42,                                 # debug randome seed. To disable >=0.
    result_precision = 3,                            # Sets the rounding for the mean values in significan_means.
    pvalue = 0.05,                                   # P-value threshold to employ for significance.
    subsampling = False,                             # To enable subsampling the data (geometri sketching).
    subsampling_log = False,                         # (mandatory) enable subsampling log1p for non log-transformed data inputs.
    subsampling_num_pc = 100,                        # Number of componets to subsample via geometric skectching (dafault: 100).
    subsampling_num_cells = 1000,                    # Number of cells to subsample (integer) (default: 1/3 of the dataset).
    separator = '|',                                 # Sets the string to employ to separate cells in the results dataframes "cellA|CellB".
    debug = False,                                   # Saves all intermediate tables employed during the analysis in pkl format.
    output_path = out_path,                          # Path to save results.
    output_suffix = None                             # Replaces the timestamp in the output files by a user defined string in the  (default: None).
    )

pvalues.head(3)
deconvoluted.head(3)
means.head(3)
